Moved InspectorsHelper out of Inspectors.scala and into InspectorsHelper.scala.
